# openapi/mobile-openapi.yaml
openapi: 3.0.3
info:
  title: TexaPay Mobile API
  version: 1.0.0
servers:
  - url: https://malika-jubilant-seriously.ngrok-free.dev
    description: Ngrok tunnel
paths:
...
  /api/mobile/transfers:
    get:
      summary: List transfers for the signed-in user
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: perPage
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransfersIndex'
  /api/mobile/transfers/{transfer}:
    get:
      summary: Get a single transfer
      parameters:
        - in: path
          name: transfer
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferSummary'
...
components:
  schemas:
...
    TransferSummary:
      type: object
      properties:
        id: { type: integer }
        status: { type: string }
        amountXaf: { type: integer }
        totalPayXaf: { type: integer }
        receiveNgnMinor: { type: integer }
        bankCode: { type: string }
        bankName: { type: string, nullable: true }
        accountNumber: { type: string }
        accountName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    TransfersIndex:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TransferSummary'
        meta:
          type: object
          properties:
            page: { type: integer }
            perPage: { type: integer }
            total: { type: integer }
            lastPage: { type: integer }
...